+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2022-09-13
+
+- 添加 GameOver Scene
+
 ## 2022-09-12
 
 - 创建 CHANGELOG